Fix serialization edge cases and remove dead/duplicate code
utils/serialization.py:
- make_json_safe: always return list for tuple input (JSON has no
  tuple type, so a stored tuple always came back as a list on a cache
  hit anyway; the distinction was never preserved).
- make_json_safe: remove a dead pd.isnull() check inside the
  pd.Timestamp branch (pd.Timestamp("NaT") is pd.NaT, so a null
  Timestamp is always caught by the earlier identity check first).
- make_json_safe: guard np.timedelta64 the same way np.datetime64
  already is, so np.timedelta64("NaT") returns None instead of a raw
  NaN float.
- stringify_special_floats: also match np.floating (e.g. np.float32),
  not just the builtin float type, so NaN/Inf in non-float64 columns
  get stringified instead of silently passing through.

Test coverage for utils/serialization.py now exercises every branch,
including the tuple case, 0-d ndarrays, datetime64/timedelta64 (both
normal and NaT), and the unhandled-type fallback. Moved the astropy
Time import to the top of the test module - it's a hard dependency
used unconditionally elsewhere in the codebase, so the try/except
ImportError guard was unnecessary.

Removed dead code from utils/dayobs.py: datetime_to_dayobs, dayobs_str,
and the dayobs2dt alias all had zero call sites.

services/jira.py: collapsed filter_tickets_with_instrument_match and
filter_tickets_without_instrument_match, which were identical apart
from a boolean polarity flip, into filter_tickets_by_instrument(...,
exclude=False). Added direct unit tests for it.

services/base_service.py: chain the caught exception with `from e` on
both re-raised HTTPExceptions for a clearer traceback.

services/exposures.py: drop a redundant float() wrap on a value that
was already a float.

main.py: drop the unused List import in favor of the builtin list[str]
already used elsewhere in the file, and remove a leftover scaffolding
comment on the CORS origins list.

+++ python/lsst/ts/logging_and_reporting/services/exposures.py
@@ -224,7 +224,7 @@ def _sum_on_sky_within_twilight(visits: pd.DataFrame, almanac_info: list[dict]) 
         "sum_visit_gap_with_filter_change": float(visit_gap[with_change_mask].sum()),
         "sum_visit_gap_without_filter_change": float(visit_gap[without_change_mask].sum()),
     }
-    return {key: round(float(total / SECONDS_IN_AN_HOUR), 2) for key, total in sums_sec.items()}
+    return {key: round(total / SECONDS_IN_AN_HOUR, 2) for key, total in sums_sec.items()}
 
 
 class ExposuresService(Service):
